Commit #1
I have the basic layout of the house all ready to go and the player
moves and the camera follows although I need to write a script for the
camera follow rather than make it a child of the player.  I also need to
set up a few walls and doors for the layout and then add some enemies
and items and UI elements.