Commit #2
After some struggles I figured out a possible solution to my lighting
problem and so I made a sprite to cover the player in darkness that will
fade once a light is lit. This will need a script with a timer to
control it. To Do: a lot, everything from the last list is basically
carried over.